Common create_mnist_module
